refactor(course-page-edit): extract block gutter alignment hook
